CT-122 블로그 게시글 페이지 리팩토링
